+++ checker-qual/src/main/java/org/checkerframework/checker/signature/qual/Identifier.java
@@ -13,11 +13,7 @@ import java.lang.annotation.Target;
  *
  * @checker_framework.manual #signature-checker Signature Checker
  */
-@SubtypeOf({
-    DotSeparatedIdentifiers.class,
-    BinaryNameWithoutPackage.class,
-    IdentifierOrPrimitiveType.class
-})
+@SubtypeOf({DotSeparatedIdentifiers.class, IdentifierOrPrimitiveType.class, InternalForm.class})
 @Documented
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.TYPE_USE, ElementType.TYPE_PARAMETER})