+++ framework/src/main/java/org/checkerframework/framework/source/SourceChecker.java
@@ -1244,7 +1244,7 @@ public abstract class SourceChecker extends AbstractTypeProcessor implements Opt
             } catch (Exception e) {
                 throw new TypeSystemError("Could not create an instance of " + subcheckerClass, e);
             }
-
+            instance.setParentChecker(this);
             immediateSubcheckers.add(instance);
             instance.setProcessingEnvironment(this.processingEnv);
             instance.treePathCacher = this.getTreePathCacher();
@@ -1252,7 +1252,6 @@ public abstract class SourceChecker extends AbstractTypeProcessor implements Opt
             instance.subcheckers = Collections.emptyList();
             instance.immediateSubcheckers =
                     instance.instantiateSubcheckers(alreadyInitializedSubcheckerMap);
-            instance.setParentChecker(this);
             alreadyInitializedSubcheckerMap.put(subcheckerClass, instance);
         }